+++ src/main/java/com/github/camilormz/finalreality/model/character/Enemy.java
@@ -1,6 +1,5 @@
 package com.github.camilormz.finalreality.model.character;
 
-import com.github.camilormz.finalreality.model.character.player.CharacterClass;
 import java.util.Objects;
 import java.util.concurrent.BlockingQueue;
 import org.jetbrains.annotations.NotNull;