+++ ChangeLog.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Version 2.0
 -----------
+- (B.2) Turning of ``PlayerCharacter`` into an abstract class ``AbstractPlayerCharacter`` and split
+        it into its subclasses, removal of the previous implementation tests.
 - (B.2) Defining of ``CharacterDomain`` as set of characteristics and behaviours of the character
         relative to the game mechanics, and restricted ``CharacterClass`` to only
         ``PlayerCharacter``.